backend/requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5

# Setup tools (required for drf-yasg)
setuptools==69.0.3

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.1

# Environment variables
python-decouple==3.8

# Password validation
argon2-cffi==23.1.0

# Image handling
Pillow==10.2.0

# Timezone support
pytz==2024.1

# API Documentation
drf-yasg==1.21.7

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Code quality
black==24.1.1
flake8==7.0.0
isort==5.13.2

# Development
ipython==8.20.0
django-extensions==3.2.3
